Redefine an action
An action, which formerly consisted of a function and a symbol to
denote the target window, is now defined in terms of
a `corefighter-action` and a payload.

This is intended for separating item data from action functions
to allow defining extra actions on modules.
The default action on an item is now called a navigation action,
which means displaying information on the item. It is included
in the module cursor, which can be retrieved from a cursor which
refers to it.

`corefighter-action` consists of a function that takes a payload
as the argument and the target window. This type may be enhanced
in the future.